requirements.txt
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2017.4.17
chardet==3.0.4
cherrypy==3.2.2
coverage==4.4.1
crispy-forms-semantic-ui==0.2.0
cycler==0.10.0
dj-database-url==0.4.2
Django==1.11.3
django-crispy-forms==1.6.1
django-google-charts==0.1.1
django-qsstats-magic==1.0.0
gunicorn==19.7.1
httplib2==0.10.3
idna==2.5
matplotlib==2.0.2
numpy==1.12.1
playsound==1.2.2
psycopg2==2.7.3
PyJWT==1.5.0
pyparsing==2.2.0
pypiwin32==220
PyQt5==5.9
PySocks==1.6.7
python-dateutil==2.6.0
python-magic==0.4.13
pyttsx==1.1
pyttsx3==2.6
pytz==2017.2
requests==2.18.1
runp==0.0.2
sip==4.19.3
six==1.10.0
twilio==6.4.2
urllib3==1.21.1
virtualenv==15.1.0
whitenoise==3.3.1
